Rename profile 'flac' to 'home'
PCM music is encoded to FLAC, but DSD music wouldn't. Indeed, 'flac'
doesn't convey the use case of the profile, but rather points to an
"implementation detail".

The names 'computer', 'pc' and 'desktop' would not fit either, because
this profile would be used on laptops and, eventually, a Kodi appliance.

What truly defines this profile or, rather, the purpose of this profile,
is that it's used at home, where storage space isn't a concern. Thus,
the best name is 'home'.